Move login & logout query to mutation
